Add key to int for app 01e6f9da-a7c3-4ae5-a9e5-595d4e2e36a6 on behalf of genesiscare
Organisation: genesiscare
Contact: ian.brittain@genesiscare.co.uk

We are beginning the integration process with the new PDS Fhir API service (we had an onboarding call today), and we would like to register our public key with you. We also ask that you host the public key for us as we do not have a way to do this. Please find the required details necessary below: